# back/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API - Agenciamento de Transferência de Jogadores",
    "description": "Documentação da API do sistema de agenciamento de transferência de jogadores para clubes. Rotas protegidas exigem um token JWT: faça login em /clubes/login ou /admin/login, copie o token e clique em \"Authorize\" abaixo, informando: Bearer SEU_TOKEN",
    "version": "1.0.0"
  },
  "basePath": "/",
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Informe: Bearer SEU_TOKEN (token retornado por /clubes/login ou /admin/login)"
    }
  },
  "paths": {
...
    "/posicoes/": {
      "get": {
        "tags": [
          "Posições"
        ],
        "summary": "Lista todas as posições",
        "description": "Retorna a lista de posições cadastradas (Goleiro, Zagueiro, Lateral, Volante, Meia, Atacante...), usadas para classificar os jogadores.",
        "responses": {
          "200": {
            "description": "Lista de posições retornada com sucesso."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Posições"
        ],
        "summary": "Cadastra uma posição",
        "description": "Cria uma nova posição, usada na classificação dos jogadores.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Atacante"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Posição cadastrada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autenticado."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/posicoes/{id}": {
      "delete": {
        "tags": [
          "Posições"
        ],
        "summary": "Exclui uma posição",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Código da posição"
          }
        ],
        "responses": {
          "200": {
            "description": "Posição excluída com sucesso."
          },
          "400": {
            "description": "Não foi possível excluir (posição em uso ou inexistente)."
          },
          "401": {
            "description": "Não autenticado."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Posições"
        ],
        "summary": "Atualiza uma posição",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Código da posição"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Meia"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Posição atualizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autenticado."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/propostas/": {
      "get": {
        "tags": [
          "Propostas"
        ],
        "summary": "Lista todas as propostas",
        "description": "Retorna todas as propostas do sistema, com os dados do jogador e do clube (sem a senha). Requer autenticação de administrador.",
        "responses": {
          "200": {
            "description": "Lista de propostas retornada com sucesso."
          },
          "401": {
            "description": "Não autenticado."
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Propostas"
        ],
        "summary": "Cria uma proposta",
        "description": "O clube autenticado (via token) faz uma proposta por um jogador. O clube da proposta é sempre o do token, não um valor enviado pelo corpo.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "valorOferta": {
                  "type": "number",
                  "example": 8000000
                },
                "mensagem": {
                  "type": "string",
                  "example": "Pago 6 milhões à vista e 2 milhões em 3x"
                },
                "jogadorId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Proposta criada com sucesso."
          },
          "400": {
            "description": "Dados inválidos ou jogador indisponível."
          },
          "401": {
            "description": "Não autenticado."
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/propostas/minhas": {
      "get": {
        "tags": [
          "Propostas"
        ],
        "summary": "Lista as propostas do clube autenticado",
        "description": "Retorna as propostas feitas pelo clube dono do token enviado (não é possível ver propostas de outro clube). Usada na tela \"Minhas Propostas\".",
        "responses": {
          "200": {
            "description": "Lista de propostas do clube retornada com sucesso."
          },
          "401": {
            "description": "Não autenticado."
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/propostas/{id}": {
      "put": {
        "tags": [
          "Propostas"
        ],
        "summary": "Responde/atualiza uma proposta",
        "description": "Usada pelo administrador para aceitar, recusar e responder uma proposta. Ao aceitar, o jogador é marcado como TRANSFERIDO, retirado dos destaques e as demais propostas pendentes por ele são recusadas automaticamente. Requer autenticação de administrador.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Código da proposta"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "resposta": {
                  "type": "string",
                  "example": "Ok. Aceitamos a proposta."
                },
                "status": {
                  "type": "string",
                  "example": "ACEITA"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proposta atualizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autenticado."
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Propostas"
        ],
        "summary": "Exclui uma proposta",
        "description": "Requer autenticação de administrador.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Código da proposta"
          }
        ],
        "responses": {
          "200": {
            "description": "Proposta excluída com sucesso."
          },
          "400": {
            "description": "Não foi possível excluir."
          },
          "401": {
            "description": "Não autenticado."
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}